% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Tidy up intensity results column names}
\usage{
clean_colnames(result)
}
\arguments{
\item{result}{a data frame with raw results from Carbon Intensity API}
}
\value{
data frame
}
\description{
Tidy up intensity results column names
}
