% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{get_british_ci}
\alias{get_british_ci}
\title{Fetch British carbon intensity data for specified time period}
\usage{
get_british_ci(date = NULL, period = NULL, start = NULL,
  end = NULL)
}
\arguments{
\item{date}{{character} A date for the intensity data}

\item{period}{{numeric} Half hour settlement period between 1-48 e.g. 42}

\item{start}{{character} A start date of the intesity data}

\item{end}{{character} An end date of the intesity data}
}
\value{
a data.frame with 1/2-hourly carbon intensity data for specified time period
}
\description{
Fetch British carbon intensity data for specified time period
}
\examples{
\dontrun{
get_british_ci()
get_british_ci(date = '2019-12-31')
get_british_ci(date = '2019-12-31', period = 48)
get_british_ci(start = '2019-01-01', end = '2019-12-31')
}
}
