% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{get_ci_by_postcode}
\alias{get_ci_by_postcode}
\title{Get Regional Carbon Intensity data for current half hour for specified postcode.}
\usage{
get_ci_by_postcode(postcode)
}
\arguments{
\item{postcode}{{character} Outward postcode i.e. RG41 or SW1 or TF8. Do not include full postcode, outward postcode only}
}
\value{
tibble
}
\description{
Get Regional Carbon Intensity data for current half hour for specified postcode.
}
\examples{
get_ci_by_postcode("SW1")
}
