% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{get_current_mix}
\alias{get_current_mix}
\title{Get generation mix for current half hour}
\usage{
get_current_mix()
}
\value{
tibble
}
\description{
Get generation mix for current half hour
}
\examples{
\dontrun{
get_current_mix()}
}
