% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{Retrieve raw data from Carbon Intensity API}
\usage{
get_data(call)
}
\arguments{
\item{call}{{character} API URL}
}
\value{
tibble
}
\description{
Retrieve raw data from Carbon Intensity API
}
