% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_factors}
\alias{get_factors}
\title{Get Carbon Intensity factors for each fuel type}
\usage{
get_factors()
}
\value{
a tibble
}
\description{
Get Carbon Intensity factors for each fuel type
}
\examples{
get_factors()
}
