% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{get_intensity}
\alias{get_intensity}
\title{Title}
\usage{
get_intensity(start, end)
}
\arguments{
\item{start}{{character} A start date of the intesity data}

\item{end}{{character} An end date of the intesity data}
}
\description{
Title
}
