% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{get_mix}
\alias{get_mix}
\title{Get generation mix for current half hour}
\usage{
get_mix(start = NULL, end = NULL)
}
\arguments{
\item{start}{{character} A start date of the intesity data}

\item{end}{{character} An end date of the intesity data}
}
\value{
tibble
}
\description{
Get generation mix for current half hour
}
\examples{
\dontrun{
start <- "2019-04-01"
end <- "2019-04-07"
get_mix(start, end)
get_mix()}
}
