% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{get_regional_ci}
\alias{get_regional_ci}
\title{Get Carbon Intensity data for current half hour for a specified GB Region}
\usage{
get_regional_ci(region = c("England", "Scotland", "Wales"))
}
\arguments{
\item{region}{{character} The name of the GB region, one of "England", "Scotland" or "Wales"}
}
\value{
a tibble
}
\description{
Get Carbon Intensity data for current half hour for a specified GB Region
}
\examples{
get_regional_ci("England")
get_regional_ci("Scotland")
get_regional_ci("Wales")
}
