% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_stats}
\alias{get_stats}
\title{Get Carbon Intensity statistics between from and to dates}
\usage{
get_stats(start, end)
}
\arguments{
\item{start}{{character} A start date of the stats data}

\item{end}{{character} An end date of the stats data}
}
\value{
tibble
}
\description{
Get Carbon Intensity statistics between from and to dates
}
\examples{
\dontrun{
start <- "2019-04-01"
end <- "2019-04-07"
get_stats(start, end)
}

}
