% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_lookup.R
\docType{data}
\name{regions_lookup}
\alias{regions_lookup}
\title{regions_lookup}
\format{A data frame with 17 rows and 2 variables:
\describe{
  \item{\code{Region ID}}{region ID to be used in intensegRid package}
  \item{\code{Shortname}}{corresponding GB region}
}}
\source{
\url{https://carbon-intensity.github.io/api-definitions/#region-list}
}
\usage{
regions_lookup
}
\description{
A lookup table of region_ids and corresponding GB regions
}
\keyword{datasets}
