% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{get_british_ci}
\alias{get_british_ci}
\title{Fetch British carbon intensity data for specified time period}
\usage{
get_british_ci(start = NULL, end = NULL)
}
\arguments{
\item{start}{{character} A start date of the intesity}

\item{end}{{character} An end date of the intesity data}
}
\value{
a data.frame with 1/2-hourly carbon intensity data for specified time period
}
\description{
Fetch British carbon intensity data for specified time period
}
\examples{
\dontrun{
get_british_ci()
get_british_ci(start = '2019-01-01', end = '2019-12-31')
}
}
