% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{get_intensity}
\alias{get_intensity}
\title{Fetch national carbon intensity data for specified time period}
\usage{
get_intensity(start, end)
}
\arguments{
\item{start}{{character} A start date of the intesity data}

\item{end}{{character} An end date of the intesity data}
}
\value{
a data.frame with 1/2-hourly carbon intensity data for specified time period
}
\description{
Fetch national carbon intensity data for specified time period
}
\examples{
\dontrun{
start <- "2019-04-01"
end <- "2019-04-07"
get_intensity(start, end)
}
}
