% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{get_regional_ci}
\alias{get_regional_ci}
\title{Get Carbon Intensity data for current half hour for a specified GB Region}
\usage{
get_regional_ci(start)

get_regional_ci(start)
}
\arguments{
\item{start}{}

\item{region}{{character} The name of the GB region, one of "England", "Scotland" or "Wales"}
}
\value{
a tibble

a tibble
}
\description{
Get Carbon Intensity data for current half hour for a specified GB Region

Get Carbon Intensity data for past 24h for GB regions
}
\examples{
get_regional("England")
get_regional("Scotland")
get_regional("Wales")
get_regional_ci()
}
